Initial commit that extracts tinyxml2 dependency to 'Serializer.h' and 'Serializer.cpp'.
