+++ Tinyxml2Serializer.h
@@ -1,8 +1,7 @@
 #pragma once
 
 #include "Settings.h"
-
-#include <tinyxml2.h>
+#include "Serializer.h"
 
 namespace qualisys_cpp_sdk {
 
@@ -13,8 +12,7 @@ namespace qualisys_cpp_sdk {
             const bool* startOnExtTrig, const bool* startOnTrigNO, const bool* startOnTrigNC, const bool* startOnTrigSoftware,
             const EProcessingActions* processingActions, const EProcessingActions* rtProcessingActions, const EProcessingActions* reprocessingActions) override;
 
-        std::string SetExtTimeBaseSettings(
-            const bool* enabled, const ESignalSource* signalSource,
+        std::string SetExtTimeBaseSettings(const bool* enabled, const ESignalSource* signalSource,
             const bool* signalModePeriodic, const unsigned int* freqMultiplier,
             const unsigned int* freqDivisor, const unsigned int* freqTolerance,
             const float* nominalFrequency, const bool* negativeEdge,
@@ -22,47 +20,37 @@ namespace qualisys_cpp_sdk {
 
         std::string SetExtTimestampSettings(const SSettingsGeneralExternalTimestamp& timestampSettings) override;
 
-        std::string SetCameraSettings(
-            const unsigned int cameraId, const ECameraMode* mode,
+        std::string SetCameraSettings(const unsigned int cameraId, const ECameraMode* mode,
             const float* markerExposure, const float* markerThreshold,
             const int* orientation) override;
 
-        std::string SetCameraVideoSettings(
-            const unsigned int cameraId, const EVideoResolution* videoResolution,
+        std::string SetCameraVideoSettings(const unsigned int cameraId, const EVideoResolution* videoResolution,
             const EVideoAspectRatio* videoAspectRatio, const unsigned int* videoFrequency,
             const float* videoExposure, const float* videoFlashTime) override;
 
-        std::string SetCameraSyncOutSettings(
-            const unsigned int cameraId, const unsigned int portNumber, const ESyncOutFreqMode* syncOutMode,
-            const unsigned int* syncOutValue, const float* syncOutDutyCycle,
-            const bool* syncOutNegativePolarity) override;
+        std::string SetCameraSyncOutSettings(const unsigned int cameraId, const unsigned int portNumber, const ESyncOutFreqMode* syncOutMode,
+            const unsigned int* syncOutValue, const float* syncOutDutyCycle, const bool* syncOutNegativePolarity) override;
 
         std::string SetCameraLensControlSettings(const unsigned int cameraId, const float focus, const float aperture) override;
+
         std::string SetCameraAutoExposureSettings(const unsigned int cameraId, const bool autoExposure, const float compensation) override;
+
         std::string SetCameraAutoWhiteBalance(const unsigned int cameraId, const bool enable) override;
-        std::string SetImageSettings(
-            const unsigned int  cameraId, const bool* enable, const CRTPacket::EImageFormat* format,
+
+        std::string SetImageSettings(const unsigned int  cameraId, const bool* enable, const CRTPacket::EImageFormat* format,
             const unsigned int* width, const unsigned int* height, const float* leftCrop,
             const float* topCrop, const float* rightCrop, const float* bottomCrop) override;
 
-        std::string SetForceSettings(
-            const unsigned int plateId, const SPoint* corner1, const SPoint* corner2,
+        std::string SetForceSettings(const unsigned int plateId, const SPoint* corner1, const SPoint* corner2,
             const SPoint* corner3, const SPoint* corner4) override;
 
         std::string Set6DOFBodySettings(const std::vector<SSettings6DOFBody>& settings6Dofs) override;
+
         std::string SetSkeletonSettings(const std::vector<SSettingsSkeletonHierarchical>& settingsSkeletons) override;
 
     private:
         std::uint32_t mMajorVersion;
         std::uint32_t mMinorVersion;
-        void AddXMLElementBool(tinyxml2::XMLElement& parentElem, const char* elementName, const bool* value, tinyxml2::XMLDocument& document, const char* trueText = "True", const char* falseText = "False");
-        void AddXMLElementBool(tinyxml2::XMLElement& parentElem, const char* elementName, const bool value, tinyxml2::XMLDocument& document, const char* trueText = "True", const char* falseText = "False");
-        void AddXMLElementInt(tinyxml2::XMLElement& parentElem, const char* elementName, const int* value, tinyxml2::XMLDocument& document);
-        void AddXMLElementUnsignedInt(tinyxml2::XMLElement& parentElem, const char* elementName, const unsigned int value, tinyxml2::XMLDocument& document);
-        void AddXMLElementUnsignedInt(tinyxml2::XMLElement& parentElem, const char* elementName, const unsigned int* value, tinyxml2::XMLDocument& document);
-        void AddXMLElementFloat(tinyxml2::XMLElement& parentElem, const char* elementName, const float* value, unsigned int decimals, tinyxml2::XMLDocument& document);
-        void AddXMLElementFloatWithTextAttribute(tinyxml2::XMLDocument& document, tinyxml2::XMLElement& parentElem, const char* elementName, const char* attributeName, const float& value, unsigned int decimals);
-        void AddXMLElementTransform(tinyxml2::XMLDocument& document, tinyxml2::XMLElement& parentElem, const std::string& name, const SPosition& position, const SRotation& rotation);
-        void AddXMLElementDOF(tinyxml2::XMLDocument& document, tinyxml2::XMLElement& parentElem, const std::string& name, const SDegreeOfFreedom& degreeOfFreedoms);
+        Serializer mSerializer;
     };
 }
